undo changes and add step to startgrasp wrapper
